refactor: 퀴즈 카테고리 상태 및 searchParams 연동을 커스텀훅으로 분리
