+++ src/components/quiz/list/QuizListContainer.tsx
@@ -10,29 +10,15 @@ import OpenQuizList from '@/components/quiz/list/OpenQuizList';
 import CloseQuizList from '@/components/quiz/list/CloseQuizList';
 import { useLectureId } from '@/hooks/queries/useLectureId';
 import TemporaryQuizList from '@/components/quiz/list/TemporaryQuizList';
+import { useQuizCategory } from '@/features/quiz/hooks/useQuizCategory';
 
 export default function QuizListContainer() {
   const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const router = useRouter();
-
-  const currentQuizCategory = searchParams.get('category') ?? '전체';
-
-  const handleQuizCategoryChange = (newCategory: string) => {
-    const params = new URLSearchParams(searchParams.toString());
-
-    if (newCategory === '전체') {
-      params.delete('category');
-    } else {
-      params.set('category', newCategory);
-    }
-
-    router.push(`${pathname}?${params.toString()}`);
-  };
+  const { currentQuizCategory, changeQuizCategory } = useQuizCategory();
 
   const QUIZ_DROPDOWN_LIST: MenuItem[] = [
-    { id: 'quiz-list-total', label: '전체', onClick: () => handleQuizCategoryChange('전체') },
-    { id: 'quiz-list-tempotaty', label: '임시저장', onClick: () => handleQuizCategoryChange('임시저장') },
+    { id: 'quiz-list-total', label: '전체', onClick: () => changeQuizCategory('전체') },
+    { id: 'quiz-list-tempotaty', label: '임시저장', onClick: () => changeQuizCategory('임시저장') },
   ];
 
   const { selectedItem: selectedQuizCategory, menuListForDropdown: quizCategoryList } = useDropdown({
